+++ lizard_map/static/lizard_map/jquery.workspace.js
@@ -89,34 +89,43 @@ function show_popup(data) {
         if (data.html && data.html.length !== 0) {
             // Generates pages with handlers. First only page 0 is visible.
             if (data.html.length === 1) {
-                // dialogContent(data.html[0]);
+                // add graph elements
                 $("#movable-dialog-content").html(data.html[0]);
             } else {
                 // Build up html with tabs.
-                html = '<ul id="popup-tabs" class="nav nav-tabs css-tabs">';
+                html = '<div id="popup-tabs"><ul>';
                 for (i = 0; i < data.html.length; i += 1) {
-                    html += '<li><a href="#">Resultaat ';
+                    html += '<li><a href="#popup-tab-' + (i + 1) + '">Resultaat ';
                     html += (i + 1) + '</a></li>';
                 }
-                html += '</ul><div class="popup-panes">';
+                html += '</ul>';
                 for (i = 0; i < data.html.length; i += 1) {
-                    html += '<div class="pane"><div class="pane-content">';
+                    html += '<div id="popup-tab-' + (i + 1) + '">';
                     html += data.html[i];
-                    html += '</div></div>';
+                    html += '</div>';
                 }
                 html += '</div>';
 
+                // add graph elements
                 $("#movable-dialog-content").html(html);
-                $("#popup-tabs").tabs("div.popup-panes > div.pane",
-                                {'effect': 'map_popup'});
+
+                $("#popup-tabs").tabs({
+                    idPrefix: 'popup-tab',
+                    selected: 0,
+                    show: function(event, ui) {
+                        resizeGraph($(ui.panel).find('.flot-graph'));
+                    }
+                });
             }
-            $("#popup-subtabs").tabs("div.popup-subpanes > div.subpane",
-                            {'effect': 'map_popup'});
+            $("#popup-subtabs").tabs({
+                idPrefix: 'popup-subtab',
+                selected: 0
+            });
             //dialogOverlay();
             $("#movable-dialog").dialog("open");
             //if (data.html.length === 1) {
                 // The tabs don't do their reload magic.
-                reloadGraphs();
+            reloadGraphs();
             //} else {
                 // Re-reload the first one.
                 //reloadLocalizedGraphs(
@@ -410,39 +419,43 @@ jQuery.fn.exists = function () {
     return this.length !== 0;
 }
 
-
-/* This wrapper is needed because the callback function of .dialog has
- * function parameters. */
-function reloadGraphsTab() {
-    reloadGraphs();
+function resizeGraph($el) {
+    if ($el) {
+        var plot = $el.data('plot');
+        if (plot) {
+            plot.resize();
+            plot.setupGrid();
+            plot.draw();
+            return true;
+        }
+    }
+    return false;
 }
 
 $(document).ready(function () {
     // Used by show_popup
     $('#content').append('<div id="movable-dialog"><div id="movable-dialog-content"></div></div>');
-    $('#movable-dialog')
-		.dialog({
-			autoOpen: false,
-			title: '',
-            resizeStop: reloadGraphsTab,
-            minWidth: 300,
-            minHeight: 300,
-            width: 800,
-            zIndex: 10000
-		})
+    $('#movable-dialog').dialog({
+		autoOpen: false,
+		title: '',
+        //resize: resizeCurrentGraph,
+        width: 800,
+        height: 480,
+        zIndex: 10000
+	});
     // Change "default" effect: reload graphs to fix layout.
-    $.tools.tabs.addEffect(
-        "map_popup",
-        function (tabIndex, done) {
+    //$.tools.tabs.addEffect(
+        //"map_popup",
+        //function (tabIndex, done) {
             // hide all panes and show the one that is clicked.
-	        this.getPanes().hide().eq(tabIndex).show();
-            done.call();
-            if ($('#movable-dialog').is(":visible")) {
+	        //this.getPanes().hide().eq(tabIndex).show();
+            //done.call();
+            //if ($('#movable-dialog').is(":visible")) {
                 // Don't reload the popup graph if the overlay isn't
                 // initialized yet.  There's no width then :-)
                 // It *does* reload twice when an overlay is already
                 // open, but we'll accept that for now.
-                reloadLocalizedGraphs(this.getPanes()[tabIndex]);
-            }
-        });
+                //reloadLocalizedGraphs(this.getPanes()[tabIndex]);
+            //}
+        //});
 });